src/requirements.txt
Django==2.1.2
django-taggit==0.23.0
Markdown==3.0.1
Pillow==5.3.0
psycopg2==2.7.5
python-decouple==3.1
pytz==2018.7
whitenoise==4.1
